+++ db/db.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/inconshreveable/log15"
-	"github.com/spf13/viper"
 	"github.com/vulsio/go-exploitdb/models"
 	"golang.org/x/xerrors"
 )
@@ -61,9 +60,32 @@ func NewDB(dbType string, dbPath string, debugSQL bool) (driver DB, locked bool,
 func newDB(dbType string) (DB, error) {
 	switch dbType {
 	case dialectSqlite3, dialectMysql, dialectPostgreSQL:
-		return &RDBDriver{name: dbType, batchSize: viper.GetInt("batch-size")}, nil
+		return &RDBDriver{name: dbType}, nil
 	case dialectRedis:
 		return &RedisDriver{name: dbType}, nil
 	}
 	return nil, fmt.Errorf("Invalid database dialect, %s", dbType)
+}
+
+// IndexChunk has a starting point and an ending point for Chunk
+type IndexChunk struct {
+	From, To int
+}
+
+func chunkSlice(length int, chunkSize int) <-chan IndexChunk {
+	ch := make(chan IndexChunk)
+
+	go func() {
+		defer close(ch)
+
+		for i := 0; i < length; i += chunkSize {
+			idx := IndexChunk{i, i + chunkSize}
+			if length < idx.To {
+				idx.To = length
+			}
+			ch <- idx
+		}
+	}()
+
+	return ch
 }
