+++ db/rdb.go
@@ -11,6 +11,7 @@ import (
 	"github.com/cheggaaa/pb/v3"
 	"github.com/inconshreveable/log15"
 	sqlite3 "github.com/mattn/go-sqlite3"
+	"github.com/spf13/viper"
 	"github.com/vulsio/go-exploitdb/config"
 	"github.com/vulsio/go-exploitdb/models"
 	"github.com/vulsio/go-exploitdb/util"
@@ -32,9 +33,8 @@ const (
 
 // RDBDriver :
 type RDBDriver struct {
-	name      string
-	conn      *gorm.DB
-	batchSize int
+	name string
+	conn *gorm.DB
 }
 
 // Name return db name
@@ -143,6 +143,8 @@ func (r *RDBDriver) deleteAndInsertExploit(conn *gorm.DB, exploitType models.Exp
 		tx.Commit()
 	}()
 
+	batchSize := viper.GetInt("batch-size")
+
 	oldIDs := []int64{}
 	result := tx.Model(&models.Exploit{}).Select("id").Where("exploit_type = ?", exploitType).Find(&oldIDs)
 	if result.Error != nil && !errors.Is(result.Error, gorm.ErrRecordNotFound) {
@@ -152,7 +154,7 @@ func (r *RDBDriver) deleteAndInsertExploit(conn *gorm.DB, exploitType models.Exp
 	if result.RowsAffected > 0 {
 		log15.Info("Deleting old Exploits")
 		bar := pb.StartNew(len(oldIDs))
-		for idx := range chunkSlice(len(oldIDs), r.batchSize) {
+		for idx := range chunkSlice(len(oldIDs), batchSize) {
 			osIDs := []int64{}
 			if err := tx.Model(&models.OffensiveSecurity{}).Select("id").Where("exploit_id IN ?", oldIDs[idx.From:idx.To]).Find(&osIDs).Error; err != nil {
 				return xerrors.Errorf("Failed to select old OffensiveSecurity: %w", err)
@@ -184,7 +186,7 @@ func (r *RDBDriver) deleteAndInsertExploit(conn *gorm.DB, exploitType models.Exp
 
 	log15.Info("Inserting new Exploits")
 	bar := pb.StartNew(len(exploits))
-	for idx := range chunkSlice(len(exploits), r.batchSize) {
+	for idx := range chunkSlice(len(exploits), batchSize) {
 		if err = tx.Create(exploits[idx.From:idx.To]).Error; err != nil {
 			return fmt.Errorf("Failed to insert. err: %s", err)
 		}
@@ -313,29 +315,6 @@ func (r *RDBDriver) GetExploitMultiByCveID(cveIDs []string) (exploits map[string
 		exploits[cveID] = r.GetExploitByCveID(cveID)
 	}
 	return exploits
-}
-
-// IndexChunk has a starting point and an ending point for Chunk
-type IndexChunk struct {
-	From, To int
-}
-
-func chunkSlice(length int, chunkSize int) <-chan IndexChunk {
-	ch := make(chan IndexChunk)
-
-	go func() {
-		defer close(ch)
-
-		for i := 0; i < length; i += chunkSize {
-			idx := IndexChunk{i, i + chunkSize}
-			if length < idx.To {
-				idx.To = length
-			}
-			ch <- idx
-		}
-	}()
-
-	return ch
 }
 
 // IsExploitModelV1 determines if the DB was created at the time of go-exploitdb Model v1
