+++ db/redis.go
@@ -227,19 +227,36 @@ func (r *RedisDriver) InsertExploit(_ models.ExploitType, exploits []models.Expl
 
 	bar := pb.StartNew(len(exploits))
 	var noCveIDExploitCount, cveIDExploitCount int
-	for _, exploit := range exploits {
+	for idx := range chunkSlice(len(exploits), viper.GetInt("batch-size")) {
 		pipe := r.conn.Pipeline()
-		bar.Increment()
+		for _, exploit := range exploits[idx.From:idx.To] {
+			j, err := json.Marshal(exploit)
+			if err != nil {
+				return fmt.Errorf("Failed to marshal json. err: %s", err)
+			}
 
-		j, err := json.Marshal(exploit)
-		if err != nil {
-			return fmt.Errorf("Failed to marshal json. err: %s", err)
-		}
+			if 0 < len(exploit.CveID) {
+				key := cveIDPrefix + exploit.CveID
+				if result := pipe.SAdd(ctx, key, string(j)); result.Err() != nil {
+					return fmt.Errorf("Failed to HSet CVE. err: %s", result.Err())
+				}
+				if expire > 0 {
+					if err := pipe.Expire(ctx, key, time.Duration(expire*uint(time.Second))).Err(); err != nil {
+						return fmt.Errorf("Failed to set Expire to Key. err: %s", err)
+					}
+				} else {
+					if err := pipe.Persist(ctx, key).Err(); err != nil {
+						return fmt.Errorf("Failed to remove the existing timeout on Key. err: %s", err)
+					}
+				}
+				cveIDExploitCount++
+			} else {
+				noCveIDExploitCount++
+			}
 
-		if 0 < len(exploit.CveID) {
-			key := cveIDPrefix + exploit.CveID
+			key := exploitDBIDPrefix + exploit.ExploitUniqueID
 			if result := pipe.SAdd(ctx, key, string(j)); result.Err() != nil {
-				return fmt.Errorf("Failed to HSet CVE. err: %s", result.Err())
+				return fmt.Errorf("Failed to HSet Exploit. err: %s", result.Err())
 			}
 			if expire > 0 {
 				if err := pipe.Expire(ctx, key, time.Duration(expire*uint(time.Second))).Err(); err != nil {
@@ -250,28 +267,11 @@ func (r *RedisDriver) InsertExploit(_ models.ExploitType, exploits []models.Expl
 					return fmt.Errorf("Failed to remove the existing timeout on Key. err: %s", err)
 				}
 			}
-			cveIDExploitCount++
-		} else {
-			noCveIDExploitCount++
 		}
-
-		key := exploitDBIDPrefix + exploit.ExploitUniqueID
-		if result := pipe.SAdd(ctx, key, string(j)); result.Err() != nil {
-			return fmt.Errorf("Failed to HSet Exploit. err: %s", result.Err())
-		}
-		if expire > 0 {
-			if err := pipe.Expire(ctx, key, time.Duration(expire*uint(time.Second))).Err(); err != nil {
-				return fmt.Errorf("Failed to set Expire to Key. err: %s", err)
-			}
-		} else {
-			if err := pipe.Persist(ctx, key).Err(); err != nil {
-				return fmt.Errorf("Failed to remove the existing timeout on Key. err: %s", err)
-			}
-		}
-
 		if _, err = pipe.Exec(ctx); err != nil {
 			return fmt.Errorf("Failed to exec pipeline. err: %s", err)
 		}
+		bar.Add(idx.To - idx.From)
 	}
 	bar.Finish()
 	log15.Info("No CveID Exploit Count", "count", noCveIDExploitCount)
